refactor: separate to package
