+++ src/main/kotlin/racingcar/RacingCarGame.kt
@@ -1,5 +1,10 @@
 package racingcar
 
+import racingcar.domain.Cars
+import racingcar.domain.RidingCount
+import racingcar.view.InputView
+import racingcar.view.OutputView
+
 class RacingCarGame {
     val inputView = InputView()
     val outputView = OutputView()
