+++ src/main/kotlin/racingcar/domain/Cars.kt
@@ -1,4 +1,6 @@
-package racingcar
+package racingcar.domain
+
+import racingcar.MovementRule
 
 class Cars(val cars: List<Car>) {
