+++ src/main/kotlin/racingcar/view/OutputView.kt
@@ -1,4 +1,7 @@
-package racingcar
+package racingcar.view
+
+import racingcar.domain.Car
+import racingcar.domain.Cars
 
 class OutputView {
