+++ src/test/kotlin/racingcar/CarsTest.kt
@@ -3,6 +3,8 @@ package racingcar
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
+import racingcar.domain.Car
+import racingcar.domain.Cars
 
 class CarsTest {
     @Test
